Fix Search Bar
